encryptor with stream index

+++ CryptZip/Compression/Token.cs
@@ -6,9 +6,9 @@
 
         public int Length { get; set; }
 
-        public byte Byte { get; set; }
+        public int Byte { get; set; }
 
-        public static Token Empty(byte b)
+        public static Token Empty(int b)
         {
             return new Token { Byte = b };
         }
